style: 질문유형 객체 이름 변경
- category -> questionCategory
