--- server/src/repositories/categoryRepository.ts
@@ -1,9 +1,0 @@
-import CategoryModel from '../models/categoryModel';
-
-import Repository from './repository';
-
-export default class CategoryRepository extends Repository {
-  constructor() {
-    super({ Model: new CategoryModel() });
-  }
-}
