--- server/src/services/category/categoryType.ts
@@ -1,10 +1,0 @@
-export default `
-type Category {
-  id: ID!
-  title: String! 
-}
-
-type Query {
-  allCategories: [Category!]!
-}
-`;
