+++ server/src/services/questionCategory/questionCategoryResolver.ts
@@ -1,0 +1,12 @@
+import QuestionCategoryRepository from '../../repositories/questionCategoryRepository';
+
+const questionCategoryRepo = new QuestionCategoryRepository();
+
+export default {
+  Query: {
+    allQuestionCategories: async () => {
+      const questionCategories = await questionCategoryRepo.findAll();
+      return questionCategories;
+    },
+  },
+};
