+++ server/src/services/questionCategory/questionCategoryType.ts
@@ -1,0 +1,10 @@
+export default `
+type QuestionCategory {
+  id: ID!
+  title: String! 
+}
+
+type Query {
+  allQuestionCategories: [QuestionCategory!]!
+}
+`;
